+++ src/misc/Enums.js
@@ -99,6 +99,7 @@ Enums.KeyAction = Utils.extend({
   CAMERA_LEFT: acc++,
   CAMERA_RESET: acc++,
   CAMERA_FOV: acc++,
+  CAMERA_PROJECTION: acc++,
   STRIFE_LEFT: acc++,
   STRIFE_RIGHT: acc++,
   STRIFE_UP: acc++,